requirements.txt
torch
torchvision
torchaudio
pytorch-lightning
omegaconf
thop
onnx
pandas
seaborn
matplotlib
scikit-learn
hydra-core
tensorboard
onnxruntime